chore(backend): separate DDB->RDS migration from normal migration
